backend/requirements.txt
asgiref==3.4.1
autopep8==1.6.0
boto3==1.20.24
botocore==1.23.24
cffi==1.15.0
cryptography==36.0.1
DateTime==4.3
Django==3.2.9
jmespath==0.10.0
pycodestyle==2.8.0
pycparser==2.21
PyJWT==1.5.3
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
uuid==1.30
zope.interface==5.4.0
asgiref==3.4.1
autopep8==1.6.0
boto3==1.20.24
botocore==1.23.24
cffi==1.15.0
cryptography==36.0.1
DateTime==4.3
Django==3.2.9
django-cors-headers==3.10.1
jmespath==0.10.0
pycodestyle==2.8.0
pycparser==2.21
PyJWT==1.5.3
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
uuid==1.30
zope.interface==5.4.0
asgiref==3.4.1
autopep8==1.6.0
boto3==1.20.24
botocore==1.23.24
cffi==1.15.0
cryptography==36.0.1
DateTime==4.3
Django==3.2.9
django-cors-headers==3.10.1
jmespath==0.10.0
pycodestyle==2.8.0
pycparser==2.21
PyJWT==1.5.3
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
uuid==1.30
zope.interface==5.4.0
asgiref==3.4.1
autopep8==1.6.0
boto3==1.20.24
botocore==1.23.24
cffi==1.15.0
cryptography==36.0.1
DateTime==4.3
Django==3.2.9
django-cors-headers==3.10.1
jmespath==0.10.0
pycodestyle==2.8.0
pycparser==2.21
PyJWT==1.5.3
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
uuid==1.30
zope.interface==5.4.0
